RepositoryFactory now enforces the use of an abstract UnitOfWorkRepositoryBase class.
