+++ Safari_UnitOfWork_Transaction_Example/Implementations/RepositoryFactory.cs
@@ -1,10 +1,11 @@
-﻿using Safari_UnitOfWork_Transaction_Example.Interfaces;
+﻿using Safari_UnitOfWork_Transaction_Example.Abstractions;
+using Safari_UnitOfWork_Transaction_Example.Interfaces;
 using System;
 using System.Data.SqlClient;
 
 namespace Safari_UnitOfWork_Transaction_Example.Implementations
 {
-    public class RepositoryFactory<TRepository> : IFactory<TRepository, SqlConnection> where TRepository : IUnitOfWorkRepository
+    public class RepositoryFactory<TRepository> : IFactory<TRepository, SqlConnection> where TRepository : UnitOfWorkRepositoryBase
     {
         public TRepository Create(SqlConnection connection)
         {
