+++ Safari_UnitOfWork_Transaction_Example/Implementations/UnitOfWorkFactory.cs
@@ -6,7 +6,6 @@ namespace Safari_UnitOfWork_Transaction_Example.Implementations
     {
         public IUnitOfWork Create()
         {
-            // Repository actually needs a abstract class
             return new UnitOfWork(
                 new RepositoryFactory<UserRepository>(),
                 new ConnectionFactory());
