+++ Safari_UnitOfWork_Transaction_Example/Interfaces/IUnitOfWorkRepository.cs
@@ -1,6 +1,8 @@
-﻿namespace Safari_UnitOfWork_Transaction_Example.Interfaces
+﻿using System;
+
+namespace Safari_UnitOfWork_Transaction_Example.Interfaces
 {
-    public interface IUnitOfWorkRepository
+    public interface IUnitOfWorkRepository : IDisposable
     {
         bool Commit();
         bool Rollback();
